Features that work: Speech works great with Live Transcription, even in background mode Alert works great along with background notifications! Text has preset phrases that works! Emoji board shows emojis for communication
New changes:
Fixes to make sure the app works for VoiceOver accessibility on all major pages
Fixed the crash in the TextView
Fixes to make sure the app works with Large font sizes
Multiple Bugs with TextEditors also fixed on Speech page

Shoutouts:
Zoe and Mofe did an amazing job testing and raising critical issues that we were able to identify and fix!
Raven for hooking up the sliders to the AlertView so now we finally can edit the NoiseLength and NoiseLevelThreshold!

Coming Soon:
Settings page that works
Emojiboard can had pictures?